requirements.txt
dj-database-url==0.5.0
Django==2.1.4
django-environ==0.4.5
mysqlclient==1.3.14
pytz==2018.7
dj-database-url==0.5.0
Django==2.1.4
django-environ==0.4.5
djangorestframework==3.9.0
mysqlclient==1.3.14
pytz==2018.7
